reworked radiative power terms and test adjust
