+++ src/thermohl/power/ieee.py
@@ -10,7 +10,6 @@ import numpy as np
 import thermohl.air as air
 import thermohl.sun as sun
 from thermohl.power.base import PowerTerm
-from thermohl.power.base import RadiativeCooling as RadiativeCooling_
 from thermohl.power.joule_heating import JouleHeating as JouleHeating_
 
 
@@ -294,5 +293,64 @@ class ConvectiveCooling(PowerTerm):
         )
 
 
-class RadiativeCooling(RadiativeCooling_):
-    pass
+class RadiativeCooling(PowerTerm):
+    """Power term for radiative cooling.
+
+    Very similar to thermohl.power.base.RadiativeCooling. Difference are in the
+    Stefan-Boltzman constant value and the celsius-kelvin conversion.
+    """
+
+    def __init__(
+            self,
+            Ta: Union[float, np.ndarray],
+            D: Union[float, np.ndarray],
+            epsilon: Union[float, np.ndarray],
+            **kwargs
+    ):
+        r"""Init with args.
+
+        Parameters
+        ----------
+        Ta : float or np.ndarray
+            Ambient temperature (C).
+        D : float or np.ndarray
+            External diameter (m).
+        epsilon : float or np.ndarray
+            Emissivity.
+
+        """
+        self.Ta = Ta
+        self.D = D
+        self.epsilon = epsilon
+
+    def value(self, T: Union[float, np.ndarray]) -> Union[float, np.ndarray]:
+        r"""Compute radiative cooling using the Stefan-Boltzmann law.
+
+        Parameters
+        ----------
+        T : float or np.ndarray
+            Conductor temperature (C).
+
+        Returns
+        -------
+        float or np.ndarray
+            Power term value (W.m\ :sup:`-1`\ ).
+
+        """
+        return 17.8 * self.epsilon * self.D * (((T + 273.) / 100.)**4 - ((self.Ta + 273.) / 100.)**4)
+
+    def derivative(self, T: Union[float, np.ndarray]) -> Union[float, np.ndarray]:
+        r"""Analytical derivative of value method.
+
+        Parameters
+        ----------
+        T : float or np.ndarray
+            Conductor temperature (C).
+
+        Returns
+        -------
+        float or np.ndarray
+            Power term derivative (W.m\ :sup:`-1`\ K\ :sup:`-1`\ ).
+
+        """
+        return 4. * 1.78E-07 * self.epsilon * self.D * T**3
